requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
pdfplumber==0.10.3
PyMuPDF==1.23.8
Pillow==10.1.0
pytesseract==0.3.10
gunicorn==21.2.0

